+++ __tests__/__fixtures__/one-module-no-beforeeach-custom-hooks-name-test.js
@@ -1,0 +1,16 @@
+import { module, test } from 'qunit';
+import { setupApplicationTest } from 'ember-qunit';
+import { click, visit } from '@ember/test-helpers';
+
+const SELECTORS = Object.freeze({
+  MOCK_SELECTOR: '[data-test-nav-bar-browse]',
+});
+
+module('Acceptance | browse acceptance test', function (adsHooksV2) {
+  setupApplicationTest(adsHooksV2);
+
+  test.only('it renders browse page', async function (assert) {
+    await visit(BROWSE_URL);
+    assert.dom(SELECTORS.MOCK_SELECTOR).exists();
+  });
+});
