+++ __tests__/__fixtures__/one-module-one-beforeeach-no-hooks-new-import-input-test.js
@@ -1,0 +1,11 @@
+import { module, test } from 'qunit';
+
+module('Unit | foo', function () {
+  some.otherThing(function () {
+    noop();
+  });
+
+  test('foo', async function (assert) {
+    assert.equal('bar', 'bar');
+  });
+});
